+++ src/main/java/faang/school/accountservice/service/AccountService.java
@@ -3,6 +3,8 @@ package faang.school.accountservice.service;
 import faang.school.accountservice.dto.AccountDto;
 import faang.school.accountservice.dto.CreateAccountDto;
 
+import java.math.BigDecimal;
+
 public interface AccountService {
 
     AccountDto getById(Long id);
@@ -17,11 +19,11 @@ public interface AccountService {
 
     AccountDto closeAccount(String accountNumber);
 
-    AccountDto withdraw(String accountNumber, double amount);
+    AccountDto withdraw(String accountNumber, BigDecimal amount);
 
-    AccountDto deposit(String accountNumber, double amount);
+    AccountDto deposit(String accountNumber, BigDecimal amount);
 
-    Double getBalance(String accountNumber);
+    BigDecimal getBalance(String accountNumber);
